Move code actions into their own package
